Split the data extensions editor domain models file
