+++ extensions/ql-vscode/src/data-extensions-editor/external-api-usage.ts
@@ -14,17 +14,3 @@ export type ExternalApiUsage = {
   supported: boolean;
   usages: Call[];
 };
-
-export type ModeledMethodType =
-  | "none"
-  | "source"
-  | "sink"
-  | "summary"
-  | "neutral";
-
-export type ModeledMethod = {
-  type: ModeledMethodType;
-  input: string;
-  output: string;
-  kind: string;
-};
