+++ extensions/ql-vscode/src/data-extensions-editor/modeled-method.ts
@@ -1,0 +1,13 @@
+export type ModeledMethodType =
+  | "none"
+  | "source"
+  | "sink"
+  | "summary"
+  | "neutral";
+
+export type ModeledMethod = {
+  type: ModeledMethodType;
+  input: string;
+  output: string;
+  kind: string;
+};
